backend/requirements.txt
# requirements.txt
django>=5.2,<=5.2 # Latest stable
djangorestframework>=3.16,<=3.16 # Latest stable
psycopg2-binary>=2.9,<=3.0 # Latest stable
gunicorn>22,<=23 # Latest stable
python-dotenv>=1.1,<1.2 # Latest stable

djangorestframework-simplejwt>=5.0,<6.0
